Auth - "fully" working authorization flow
You can now search for artists and search again without any errors.
The method of authenticating is somewhat messy and probably not
the intended method. The user is essentially signing in on every
artist search. There doenst seem to be any issues so far, other
than maybe a slight delay. This code does not include the related
artists or images.