styling and some error checking
Added some page styling and error checking on the server side.
If an artist has less than 20 related artists, the loop only iterates
that amount of times. This still has an issue with very small
artists who have no related artists since the loop is never entered
and their info is never sent. Also, small artists are reporting
undefined genre. I tried catching it and setting the genre to
Unknown, but it wasnt working. Artists with no picture are getting
a default image now. The graph is ready to be placed next to the
artist result when ready. You can click the image to go to their page.