related artist chart added
The chart has a few bugs. The text is not centered next to the
image. Also, the page has a lot of dead space under. Preferably,
the page would have no scrolling. Also added hrefs for the image
and artist name. The artist name is meant to launch a new search
for that artist, but its broken right now.